# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "High-performance microservice for URL pattern matching and CSS/JS asset injection in Web-to-PDF pipelines",
        "title": "Asset Injector Microservice API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/resolve": {
            "post": {
                "description": "Matches a URL against configured rules and returns the most specific CSS/JS assets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resolution"
                ],
                "summary": "Resolve URL pattern to CSS/JS assets",
                "parameters": [
                    {
                        "description": "URL to resolve",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully resolved URL",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.ResolveResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/rules": {
            "get": {
                "description": "Retrieves all configured URL matching rules",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rules"
                ],
                "summary": "List all rules",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved rules",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.RuleListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new URL matching rule or updates an existing one",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rules"
                ],
                "summary": "Create or update a rule",
                "parameters": [
                    {
                        "description": "Rule to create or update",
                        "name": "rule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Rule"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created rule",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "rule": {
                                                    "$ref": "#/definitions/domain.Rule"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/rules/{id}": {
            "delete": {
                "description": "Deletes a URL matching rule by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rules"
                ],
                "summary": "Delete a rule",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted rule",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                },
                                                "rule_id": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Rule not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "description": "Standard error response format",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_FAILED"
                },
                "details": {},
                "message": {
                    "type": "string",
                    "example": "Invalid input provided"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "api.ResolveRequest": {
            "description": "Request payload for URL pattern resolution",
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://example.com/page"
                }
            }
        },
        "api.ResolveResponse": {
            "description": "Response payload containing matched CSS/JS assets",
            "type": "object",
            "properties": {
                "cache_hit": {
                    "type": "boolean",
                    "example": false
                },
                "css": {
                    "type": "string",
                    "example": ".banner { display: none; }"
                },
                "js": {
                    "type": "string",
                    "example": "document.querySelector('.popup').remove();"
                },
                "rule_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
        "api.RuleListResponse": {
            "description": "Response containing list of rules",
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 5
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Rule"
                    }
                }
            }
        },
        "api.SuccessResponse": {
            "description": "Standard success response format",
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "domain.Rule": {
            "description": "URL pattern matching rule configuration",
            "type": "object",
            "required": [
                "id",
                "pattern",
                "type"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                },
                "css": {
                    "description": "100KB limit",
                    "type": "string",
                    "maxLength": 102400,
                    "example": ".banner { display: none; }"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "js": {
                    "description": "100KB limit",
                    "type": "string",
                    "maxLength": 102400,
                    "example": "document.querySelector('.popup').remove();"
                },
                "pattern": {
                    "type": "string",
                    "maxLength": 2048,
                    "minLength": 1,
                    "example": "https://example.com/*"
                },
                "priority": {
                    "type": "integer",
                    "maximum": 10000,
                    "minimum": 0,
                    "example": 1500
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "exact",
                        "regex",
                        "wildcard"
                    ],
                    "example": "exact"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}